added yesno parser for module arguments

+++ joule/daemon/worker.py
@@ -141,6 +141,8 @@ class Worker:
             return
         # close pipe connections with module
         await self._close_npipes()
+        if(self.process is None):
+            return        
         self.process.terminate()
         try:
             await asyncio.wait_for(self.process.wait(),
